+++ CHANGELOG.md
@@ -55,6 +55,7 @@ This version introduce major changes in the tmp and rc file.
  + Add licence warning.
  + Remove core\_extended dependency, incorporating some code from the library
    directly in the program, and using Textutils and Re2 library instead.
+ + Display debugging information before each message.
  + TODO XXX Add basic signal handling (`--signals`), to relaunch when doing
    ctrl-C. See issue #14 for known problems.